requirements.txt
streamlit
google-cloud-aiplatform
google-generativeai
Pillow
numpy
